Refactor installer to reduce pattern duplication
